web/requirements.txt
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
charset-normalizer==2.0.9
Eel==0.14.0
emoji==1.6.1
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
pyparsing==3.0.6
requests==2.26.0
urllib3==1.26.7
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
filestack-python==3.5.0
